+++ src/general/tools/async_tool/index.ts
@@ -1,0 +1,3 @@
+export function TestFunction() {
+    console.log('async-tool package test');
+}